Move the Coverage Atlas and near-dupe algorithms out of vtscore/state/
Two of the three modules named in #3391 are pure algorithms that reference
neither DatasetContext nor _state_lock, and only sat in state/ because the
wiring that drives them does:

- vtscore/state/coverage_atlas.py -> vtscore/coverage/atlas.py (new package)
- vtscore/state/near_dupes.py     -> vtscore/media/near_dupes.py

Moving the atlas out also dissolves the state/coverage.py vs
state/coverage_atlas.py near-homograph the issue flagged, with no rename
needed: the wiring keeps the name, and the algorithm is now a package of its
own that says in its docstring that it holds no state and takes no lock.

sort_results_cache.py deliberately stays in state/. The issue described it as
a generic LRU, but the module ends with a process-global mutable singleton
guarded by its own threading.Lock, and both its docstring and the package docs
call it "process-global". Having its own lock rather than the shared one does
not make it not-state; moving it to utils/ would have made the layout less
accurate, not more.

Both old module paths remain as aliases that re-export the new location and
raise DeprecationWarning on import, with a PEP 562 __getattr__ forwarding
anything not explicitly re-exported so module privates keep resolving. These
are documented public vtscore paths, so out-of-tree code may import them.
The vtscore.state / vtsearch.state re-exports are untouched.

Docs follow the code: a new vtscore/docs/packages/coverage.md owns the
structure, state.md keeps the wiring API and points at it, and the
near-duplicate section moves to media.md.

Closes #3391
