+++ docs/plans/coverage-atlas.md
@@ -2,7 +2,7 @@
 
 **Status:** design writeup + first slice shipped. This document is the output of a first-principles design discussion ("invent the perfect density structure, don't adapt the existing one"). It scopes the science; the sections below remain the spec for the unshipped parts.
 
-**Background — what already exists:** the Diversity Tree has been migrated to a `CoverageAtlas` (`vtscore/state/coverage_atlas.py`): the same hierarchical k-means partition, now over mean-centered renormalized embeddings (§4.1's cone fix), keeping per-node per-class evidence counts (`n_pos`/`n_neg`), mean-direction moments (`mu`, `rbar`), and 21-point leave-one-out typicality quantiles. Q1 typicality is live (path-averaged p-values, calibration gated on node size ≥ 20 and `rbar` ≥ 0.1 — the root is degenerate by construction after centering). The autopilot's diversity step samples mass-first (siblings stored largest-first) and picks the node's most typical element when unscored. The atlas serializes into the dataset pickle (`"coverage_atlas"` cache key), and an in-session domain-shift report exists at `GET /api/datasets/registry/<id>/domain-shift` (active dataset's items vs `<id>`'s atlas; refuses on embedder mismatch) — now surfaced in the Find Stats modal's **Training-domain overlap** section.
+**Background — what already exists:** the Diversity Tree has been migrated to a `CoverageAtlas` (`vtscore/coverage/atlas.py`): the same hierarchical k-means partition, now over mean-centered renormalized embeddings (§4.1's cone fix), keeping per-node per-class evidence counts (`n_pos`/`n_neg`), mean-direction moments (`mu`, `rbar`), and 21-point leave-one-out typicality quantiles. Q1 typicality is live (path-averaged p-values, calibration gated on node size ≥ 20 and `rbar` ≥ 0.1 — the root is degenerate by construction after centering). The autopilot's diversity step samples mass-first (siblings stored largest-first) and picks the node's most typical element when unscored. The atlas serializes into the dataset pickle (`"coverage_atlas"` cache key), and an in-session domain-shift report exists at `GET /api/datasets/registry/<id>/domain-shift` (active dataset's items vs `<id>`'s atlas; refuses on embedder mismatch) — now surfaced in the Find Stats modal's **Training-domain overlap** section.
 
 The **v0 labelset-kNN evidence signals** (§6.1's `D` and `TS`) have also shipped — the *cross-user* decision-support signal that needs no reference dataset, since it reads only the detector's own labelset (re-embedded in memory at load). `vtscore/detectors/evidence_coverage.py` computes per-item conformal support p-values `D` (calibrated against the labelset's leave-one-out same-class distances) and trust-score ratios `TS`; `GET /api/find/evidence-coverage` summarises them over a scored Find run (fraction in an evidence vacuum for the predicted class, z-score, low-trust share) and the Find Stats modal renders an **Evidence coverage** chip beside Training-domain overlap. This is the piece of the design that fires in the true handoff where haystackA never resolves.
