+++ vtscore/state/__init__.py
@@ -113,7 +113,7 @@ from vtscore.state.media_lookup import (  # noqa: F401
 )
 
 # Re-export near-duplicate collapsing ---------------------------------------
-from vtscore.state.near_dupes import (  # noqa: F401
+from vtscore.media.near_dupes import (  # noqa: F401
     collapse_near_duplicates,
     phash_image,
     simhash_text,
